feat(light-object-parser): :art: create Runtime object
