+++ src/main/java/com/sevenmap/core/README.md
