+++ src/main/java/com/sevenmap/core/Runtime.java
@@ -1,0 +1,13 @@
+package com.sevenmap.core;
+
+import org.apache.commons.cli.CommandLine;
+
+public class Runtime {
+
+  private CommandLine cl;
+
+  public void load(CommandLine cl) {
+    this.cl = cl;
+
+  }
+}
