+++ src/main/java/com/sevenmap/core/UI/UI.java
@@ -1,0 +1,12 @@
+package com.sevenmap.core.UI;
+
+import com.sevenmap.core.Loadable;
+
+import org.apache.commons.cli.CommandLine;
+
+public class UI extends Loadable {
+
+  public void load(CommandLine cl) {
+
+  }
+}
